ajustes, mejoras y correcciones en todas las migraciones de la aplicación

+++ database/migrations/2018_04_28_022452_add_fields_to_permissions_table.php
@@ -31,8 +31,12 @@ class AddFieldsToPermissionsTable extends Migration
     public function down()
     {
         Schema::table('permissions', function (Blueprint $table) {
-            $table->dropColumn('model_prefix');
-            $table->dropColumn('slug_alt');
+            if (Schema::hasColumn('permissions', 'model_prefix')) {
+                $table->dropColumn('model_prefix');
+            }
+            if (Schema::hasColumn('permissions', 'slug_alt')) {
+                $table->dropColumn('slug_alt');
+            }
         });
     }
 }
