+++ database/migrations/2018_09_12_144844_add_field_short_description_to_permissions_table.php
@@ -28,7 +28,9 @@ class AddFieldShortDescriptionToPermissionsTable extends Migration
     public function down()
     {
         Schema::table('permissions', function (Blueprint $table) {
-            $table->dropColumn('short_description');
+            if (Schema::hasColumn('permissions', 'short_description')) {
+                $table->dropColumn('short_description');
+            }
         });
     }
 }
