+++ database/migrations/2018_09_28_183000_add_citiable_morphs_field_to_cities_table.php
@@ -28,8 +28,10 @@ class AddCitiableMorphsFieldToCitiesTable extends Migration
     public function down()
     {
         Schema::table('cities', function (Blueprint $table) {
-            $table->dropColumn('citiable_id');
-            $table->dropColumn('citiable_type');
+            if (Schema::hasColumns('cities', ['citiable_id', 'ciatiable_type'])) {
+                $table->dropColumn('citiable_id');
+                $table->dropColumn('citiable_type');
+            }
         });
     }
 }
