+++ database/migrations/2019_02_06_221844_add_field_action_to_document_status_table.php
@@ -31,8 +31,10 @@ class AddFieldActionToDocumentStatusTable extends Migration
      */
     public function down()
     {
-        Schema::table('document_status', function (Blueprint $table) {
-            $table->dropColumn('action');
-        });
+        if (Schema::hasColumn('document_status', 'action')) {
+            Schema::table('document_status', function (Blueprint $table) {
+                $table->dropColumn('action');
+            });
+        }
     }
 }
