+++ database/migrations/2019_03_19_145314_add_field_type_to_code_settings_table.php
@@ -42,8 +42,10 @@ class AddFieldTypeToCodeSettingsTable extends Migration
      */
     public function down()
     {
-        Schema::table('code_settings', function (Blueprint $table) {
-            $table->dropColumn('type');
-        });
+        if (Schema::hasColumn('code_settings', 'type')) {
+            Schema::table('code_settings', function (Blueprint $table) {
+                $table->dropColumn('type');
+            });
+        }
     }
 }
