+++ modules/Payroll/Database/Migrations/2019_05_31_142055_create_payroll_professional_informations_table.php
@@ -25,55 +25,48 @@ class CreatePayrollProfessionalInformationsTable extends Migration
      */
     public function up()
     {
-        Schema::create('payroll_professional_informations', function (Blueprint $table) {
-            $table->bigIncrements('id');
+        if (!Schema::hasTable('payroll_professional_informations')) {
+            Schema::create('payroll_professional_informations', function (Blueprint $table) {
+                $table->bigIncrements('id');
 
-            $table->bigInteger('payroll_instruction_degree_id')->unsigned()
-                  ->comment('identificador del grado de instrucción que pertenece a la información profesional');
-            $table->foreign(
-                'payroll_instruction_degree_id',
-                'payroll_professional_informations_instruction_degree_fk'
-            )->references('id')->on('payroll_instruction_degrees')->onDelete('restrict')->onUpdate('cascade');
+                $table->unsignedBigInteger('payroll_instruction_degree_id')->comment(
+                    'identificador del grado de instrucción que pertenece a la información profesional'
+                );
+                $table->foreign(
+                    'payroll_instruction_degree_id',
+                    'payroll_professional_informations_instruction_degree_fk'
+                )->references('id')->on('payroll_instruction_degrees')->onDelete('restrict')->onUpdate('cascade');
 
-            $table->bigInteger('profession_id')->unsigned()->nullable()
-                  ->comment('identificador de la profesión que pertenecen a la información profesional');
-            $table->foreign('profession_id')->references('id')->on('professions')
-                  ->onDelete('restrict')->onUpdate('cascade');
 
-            $table->string('instruction_degree_name', 100)
-                  ->nullable()
-                  ->comment('Nombre en caso de elegir en Grado de Instrucción: Especialización, Maestría, Doctorado');
+                $table->foreignId('profession_id')->nullable()->constrained()
+                      ->onDelete('restrict')->onUpdate('cascade');
 
-            $table->boolean('is_student')->default(false)->comment('Establece si el trabajdor es estudiante o no');
+                $table->string('instruction_degree_name', 100)->nullable()->comment(
+                    'Nombre en caso de elegir en Grado de Instrucción: Especialización, Maestría, Doctorado'
+                );
 
-            $table->bigInteger('payroll_study_type_id')->unsigned()->nullable()
-                  ->comment('identificador del tipo de estudio que pertenecen a la información profesional');
-            $table->foreign('payroll_study_type_id')->references('id')->on('payroll_study_types')
-                  ->onDelete('restrict')->onUpdate('cascade');
+                $table->boolean('is_student')->default(false)->comment(
+                    'Establece si el trabajdor es estudiante o no'
+                );
 
-            $table->text('study_program_name')->nullable()->comment('Nombre del programa de estudio');
-            $table->text('class_schedule')->nullable()->comment('Descripción de la cuenta');
+                $table->foreignId('payroll_study_type_id')->nullable()->constrained()
+                      ->onDelete('restrict')->onUpdate('cascade');
 
-            $table->bigInteger('payroll_language_id')->unsigned()
-                  ->comment('identificador del idioma que pertenece a la información profesional');
-            $table->foreign('payroll_language_id')->references('id')->on('payroll_languages')
-                  ->onDelete('restrict')->onUpdate('cascade');
+                $table->text('study_program_name')->nullable()->comment('Nombre del programa de estudio');
+                $table->text('class_schedule')->nullable()->comment('Descripción de la cuenta');
 
-            $table->bigInteger('payroll_language_level_id')->unsigned()
-                  ->comment('identificador de nivel del idioma que pertenece a la información profesional');
-            $table->foreign(
-                'payroll_language_level_id',
-                'payroll_professional_informations_language_level_pk'
-            )->references('id')->on('payroll_language_levels')->onDelete('restrict')->onUpdate('cascade');
+                $table->foreignId('payroll_language_id')->constrained()->onDelete('restrict')->onUpdate('cascade');
 
-            $table->bigInteger('payroll_staff_id')->unsigned()->unique()
-                  ->comment('identificador del trabajador que pertenece a la información profesional');
-            $table->foreign('payroll_staff_id')->references('id')->on('payroll_staffs')
-                  ->onDelete('restrict')->onUpdate('cascade');
+                $table->foreignId('payroll_language_level_id')->constrained()
+                      ->onDelete('restrict')->onUpdate('cascade');
 
-            $table->timestamps();
-            $table->softDeletes()->comment('Fecha y hora en la que el registro fue eliminado');
-        });
+                $table->foreignId('payroll_staff_id')->unique()->constrained()
+                      ->onDelete('restrict')->onUpdate('cascade');
+
+                $table->timestamps();
+                $table->softDeletes()->comment('Fecha y hora en la que el registro fue eliminado');
+            });
+        }
     }
 
     /**
