+++ modules/Payroll/Database/Migrations/2019_10_16_133624_delete_fields_to_payroll_professional_informations_table.php
@@ -26,10 +26,17 @@ class DeleteFieldsToPayrollProfessionalInformationsTable extends Migration
     public function up()
     {
         Schema::table('payroll_professional_informations', function (Blueprint $table) {
-            $table->dropForeign('payroll_professional_informations_payroll_language_id_foreign');
-            $table->dropColumn('payroll_language_id');
-            $table->dropForeign('payroll_professional_informations_language_level_pk');
-            $table->dropColumn('payroll_language_level_id');
+            Schema::disableForeignKeyConstraints();
+
+            if (Schema::hasColumn('payroll_professional_informations', 'payroll_language_id')) {
+                $table->dropForeign('payroll_professional_informations_payroll_language_id_foreign');
+                $table->dropColumn('payroll_language_id');
+            }
+            if (Schema::hasColumn('payroll_professional_informations', 'payroll_language_level_id')) {
+                $table->dropColumn('payroll_language_level_id');
+            }
+
+            Schema::enableForeignKeyConstraints();
         });
     }
 
@@ -42,15 +49,25 @@ class DeleteFieldsToPayrollProfessionalInformationsTable extends Migration
     public function down()
     {
         Schema::table('payroll_professional_informations', function (Blueprint $table) {
-            $table->bigInteger('payroll_language_id')->unsigned()->nullable()
-                  ->comment('identificador del idioma que pertenecen a la información profesional');
-            $table->foreign('payroll_language_id')->references('id')->on('payroll_languages')
-                  ->onDelete('restrict')->onUpdate('cascade');
+            if (!Schema::hasColumn('payroll_professional_informations', 'payroll_language_id')) {
+                $table->unsignedBigInteger('payroll_language_id')->nullable()->comment(
+                    'identificador del idioma que pertenecen a la información profesional'
+                );
+                $table->foreign(
+                    'payroll_language_id',
+                    'payroll_professional_informations_language_fk'
+                )->references('id')->on('payroll_languages')->onDelete('restrict')->onUpdate('cascade');
+            }
 
-            $table->bigInteger('payroll_language_level_id')->unsigned()->nullable()
-                  ->comment('identificador del mivel de idioma que pertenecen a la información profesional');
-            $table->foreign('payroll_language_level_id')->references('id')->on('payroll_language_levels')
-                  ->onDelete('restrict')->onUpdate('cascade');
+            if (!Schema::hasColumn('payroll_professional_informations', 'payroll_language_level_id')) {
+                $table->unsignedBigInteger('payroll_language_level_id')->nullable()->comment(
+                    'identificador del mivel de idioma que pertenecen a la información profesional'
+                );
+                $table->foreign(
+                    'payroll_language_level_id',
+                    'payroll_professional_informations_level_fk'
+                )->references('id')->on('payroll_language_levels')->onDelete('restrict')->onUpdate('cascade');
+            }
         });
     }
 }
