+++ modules/Purchase/Database/Migrations/2020_03_12_153023_create_purchase_pivot_models_to_requirement_items_table.php
@@ -13,24 +13,29 @@ class CreatePurchasePivotModelsToRequirementItemsTable extends Migration
      */
     public function up()
     {
-        Schema::create('purchase_pivot_models_to_requirement_items', function (Blueprint $table) {
-            $table->bigIncrements('id');
+        if (!Schema::hasTable('purchase_pivot_models_to_requirement_items')) {
+            Schema::create('purchase_pivot_models_to_requirement_items', function (Blueprint $table) {
+                $table->bigIncrements('id');
 
-            $table->morphs('relatable', 'purchase_pivot_models_to_requirement_items_relatable_index');
+                $table->morphs('relatable', 'purchase_pivot_models_to_requirement_items_relatable_index');
 
-            $table->bigInteger('purchase_requirement_item_id')->unsigned()
-                      ->comment('Llave foranea a un producto en un requerimiento');
-            $table->foreign(
-                'purchase_requirement_item_id',
-                'purchase_pivot_models_to_requirement_items_requirements_fk'
-            )->references('id')->on('purchase_requirement_items')->onDelete('cascade')->onUpdate('cascade');
+                $table->unsignedBigInteger('purchase_requirement_item_id')->comment(
+                    'Llave foranea a un producto en un requerimiento'
+                );
+                $table->foreign(
+                    'purchase_requirement_item_id',
+                    'purchase_pivot_models_to_requirement_items_requirements_fk'
+                )->references('id')->on('purchase_requirement_items')->onDelete('cascade')->onUpdate('cascade');
 
-            $table->float('unit_price', 10, 10)->nullable()
-                          ->comment('Precio unitario del producto o servicio. asignado en orden de compra');
 
-            $table->timestamps();
-            $table->softDeletes()->comment('Fecha y hora en la que el registro fue eliminado');
-        });
+                $table->float('unit_price', 10, 10)->nullable()->comment(
+                    'Precio unitario del producto o servicio. asignado en orden de compra'
+                );
+
+                $table->timestamps();
+                $table->softDeletes()->comment('Fecha y hora en la que el registro fue eliminado');
+            });
+        }
     }
 
     /**
